Split DemoData Into a File
